improve redux store to accommodate create token action
