+++ src/actions/types.js
@@ -1,0 +1,15 @@
+const CREATE_TOKEN = 'CREATE_TOKEN';
+const REMOVE_TOKEN = 'REMOVE_TOKEN';
+const CREATE_CONSULTANCIES_LIST = 'CREATE_CONSULTANCIES_LIST';
+const CREATE_APPOINTMENTS_LIST = 'CREATE_APPOINTMENTS_LIST';
+const SET_APPOINTMENT = 'SET_APPOINTMENT';
+const REMOVE_APPOINTMENT = 'REMOVE_APPOINTMENT';
+
+export {
+  CREATE_TOKEN,
+  REMOVE_TOKEN,
+  CREATE_CONSULTANCIES_LIST,
+  CREATE_APPOINTMENTS_LIST,
+  SET_APPOINTMENT,
+  REMOVE_APPOINTMENT,
+};
