+++ src/components/App.js
@@ -5,6 +5,7 @@ import {
 } from 'react-router-dom';
 import NavBar from '../containers/NavBar';
 import { SignUp } from '../containers/SignUp';
+import signUpRequest from '../requests/signUpRequest';
 
 function App() {
   return (
@@ -14,7 +15,7 @@ function App() {
         <NavBar />
         <Switch>
           <Route exact path="/signup" component={SignUp}>
-            <SignUp />
+            <SignUp createToken={signUpRequest} />
           </Route>
         </Switch>
       </div>
