+++ src/containers/SignUp.js
@@ -1,13 +1,11 @@
-import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { Link } from 'react-router-dom';
 import SignUpForm from '../components/signUpForm';
 import signUpRequest from '../requests/signUpRequest';
 import { createToken } from '../actions/index';
 
-export const SignUp = ({ createToken }) => {
+export const SignUp = () => {
   const handleSubmit = data => {
-    console.log(createToken);
     signUpRequest(createToken, data);
   };
 
@@ -20,12 +18,4 @@ export const SignUp = ({ createToken }) => {
   );
 };
 
-SignUp.propTypes = {
-  createToken: PropTypes.func.isRequired,
-};
-
-const mapDispatchToProp = {
-  createToken,
-};
-
-export default connect(null, mapDispatchToProp)(SignUp);
+export default connect(null, { createToken })(SignUp);
