+++ src/services/appointmentapi.js
@@ -1,4 +1,4 @@
-export const BASE = 'https://arcane-taiga-86251.herokuapp.com';
+export const BASE = 'http://localhost:9000/';
 export const CONSULTANCIES = 'consultancies';
 export const APPOINTMENTS = 'appointments';
 export const SIGNUP = 'signup';
